Create ReactContent component
